parser: split into expr.rs
